requirements.txt
aiohttp>=3.8.0,<4.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
pyyaml>=6.0.0
requests>=2.31.0
colorama>=0.4.6
